+++ include/SW_Site.h
@@ -353,7 +353,6 @@ void echo_inputs(
 /* these used to be in Layers */
 void SW_LYR_read(
     SW_SOIL_RUN_INPUTS *SW_SoilRunIn,
-    LyrIndex *n_evap_lyrs,
     LyrIndex *n_layers,
     char *txtInFiles[],
     LOG_INFO *LogInfo