+++ src/SW_Domain.c
@@ -146,7 +146,6 @@ void SW_DOM_construct(size_t rng_seed, SW_DOMAIN *SW_Domain) {
 #endif
 
     SW_Domain->nMaxSoilLayers = 0;
-    SW_Domain->nMaxEvapLayers = 0;
     SW_Domain->hasConsistentSoilLayerDepths = swFALSE;
     memset(
         &SW_Domain->depthsAllSoilLayers,
@@ -679,8 +678,9 @@ void SW_DOM_deconstruct(SW_DOMAIN *SW_Domain) {
 /**
 @brief Identify soil profile information across simulation domain
 
-nc-mode uses the same vertical size for every soil nc-output file, i.e.,
-it sets nMaxEvapLayers to nMaxSoilLayers.
+Text-mode SOILWAT2 before v8.4.0 produced output of soil evaporation only
+for those soil layers with potential evaporation coefficients larger than 0.
+Since v8.4.0, all modes produce output for each soil layer.
 
 @param[in] SW_netCDFIn Constant netCDF input file information
 @param[in] SW_PathInputs
@@ -689,13 +689,9 @@ it sets nMaxEvapLayers to nMaxSoilLayers.
     (though potentially variable number of soil layers)
 @param[out] nMaxSoilLayers Largest number of soil layers across
     simulation domain
-@param[out] nMaxEvapLayers Largest number of soil layers from which
-    bare-soil evaporation may extract water across simulation domain
 @param[out] depthsAllSoilLayers Lower soil layer depths [cm] if
     consistent across simulation domain
 @param[in] default_n_layers Default number of soil layer
-@param[in] default_n_evap_lyrs Default number of soil layer used for
-    bare-soil evaporation (only used in text-mode)
 @param[in] default_depths Default values of soil layer depths [cm]
 @param[out] LogInfo Holds information on warnings and errors
 */
@@ -704,10 +700,8 @@ void SW_DOM_soilProfile(
     SW_PATH_INPUTS *SW_PathInputs,
     Bool hasConsistentSoilLayerDepths,
     LyrIndex *nMaxSoilLayers,
-    LyrIndex *nMaxEvapLayers,
     double depthsAllSoilLayers[],
     LyrIndex default_n_layers,
-    LyrIndex default_n_evap_lyrs,
     const double default_depths[],
     LOG_INFO *LogInfo
 ) {
@@ -718,7 +712,6 @@ void SW_DOM_soilProfile(
             SW_netCDFIn,
             hasConsistentSoilLayerDepths,
             nMaxSoilLayers,
-            nMaxEvapLayers,
             depthsAllSoilLayers,
             SW_PathInputs->numSoilVarLyrs,
             default_n_layers,
@@ -726,12 +719,6 @@ void SW_DOM_soilProfile(
             LogInfo
         );
     } else {
-        /* nc-mode produces soil output across all soil layers */
-        *nMaxEvapLayers = default_n_layers;
-
-#else // !SWNETCDF
-    /* text-mode produces soil evaporation output only for evap layers */
-    *nMaxEvapLayers = default_n_evap_lyrs;
 #endif
 
         // Assume default/template values are consistent
@@ -744,12 +731,10 @@ void SW_DOM_soilProfile(
 
 #if defined(SWNETCDF)
     }
-#endif // !SWNETCDF
+#endif
 
     /* Cast unused variables to void to silence the compiler */
-#if defined(SWNETCDF)
-    (void) default_n_evap_lyrs;
-#else
+#if !defined(SWNETCDF)
     (void) SW_netCDFIn;
     (void) SW_PathInputs;
     (void) hasConsistentSoilLayerDepths;