+++ tests/gtests/sw_testhelpers.cc
@@ -317,7 +317,6 @@ int setup_testGlobalSoilwatTemplate() {
 
     SW_OUT_setup_output(
         template_SW_Run.RunIn.SiteRunIn.n_layers,
-        template_SW_Run.SiteSim.n_evap_lyrs,
         template_SW_Run.VegEstabIn.count,
         template_SW_Run.VegEstabIn.parms,
         &template_SW_Domain.OutDom,